Create LICENSE
This commit adds the MIT License to the project. The license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, and to permit persons to whom the software is furnished to do so, subject to the conditions stated in the license. The software is provided "as is", without warranty of any kind.